Refactor modem stuff out of main.c to modem.c
