+++ misc.c
@@ -993,3 +993,5 @@ char * misc_hayes_ATQ_to_str(int dtr)
     }
 
 }
+
+
